+++ stac_fastapi/extensions/stac_fastapi/extensions/core/fields/__init__.py
@@ -1,5 +1,5 @@
 """Fields extension module."""
 
-from .fields import FieldsExtension
+from .fields import FieldsConformanceClasses, FieldsExtension
 
-__all__ = ["FieldsExtension"]
+__all__ = ["FieldsExtension", "FieldsConformanceClasses"]