+++ stac_fastapi/extensions/stac_fastapi/extensions/core/filter/__init__.py
@@ -1,5 +1,17 @@
 """Filter extension module."""
 
-from .filter import FilterExtension
+from .filter import (
+    CollectionSearchFilterExtension,
+    FilterConformanceClasses,
+    FilterExtension,
+    ItemCollectionFilterExtension,
+    SearchFilterExtension,
+)
 
-__all__ = ["FilterExtension"]
+__all__ = [
+    "FilterConformanceClasses",
+    "FilterExtension",
+    "SearchFilterExtension",
+    "ItemCollectionFilterExtension",
+    "CollectionSearchFilterExtension",
+]