+++ stac_fastapi/extensions/stac_fastapi/extensions/core/query/__init__.py
@@ -1,5 +1,5 @@
 """Query extension module."""
 
-from .query import QueryExtension
+from .query import QueryConformanceClasses, QueryExtension
 
-__all__ = ["QueryExtension"]
+__all__ = ["QueryExtension", "QueryConformanceClasses"]